Moved InitializeWithAvailableUILocales to SIL.Windows.Forms
